requirements_api.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.5
requests>=2.31
numpy>=1.26
Pillow>=10.3
rembg>=2.0.66
opencv-python-headless>=4.10.0.84
trimesh>=4.4.0
scikit-image>=0.25.0
PyYAML>=6.0
